sharkle/requirements.txt
asgiref==3.4.1
Django==3.2.6
mysqlclient==2.0.1
pytz==2021.1
sqlparse==0.4.1
django-debug-toolbar==2.2
djangorestframework==3.11.1
python-dotenv
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.11.0
black==22.1.0
pre-commit==2.17.0
factory-boy==3.2.1
boto3==1.21.18
watchtower

Pillow
boto3
django-storages